Refactor(helm): extract manifest parsing into manifest.go
Move parseManifestResources, isClusterScopedGVK,
  isClusterScopedKindStaticFallback, isTestResource, cleanResource,
  and orderResources out of helm.go.

No behavior change. All 240 unit specs remain green.
